requirements.txt
alembic==1.4.2
basilica==0.2.8
certifi==2019.11.28
chardet==3.0.4
click==7.1.1
Flask==1.1.1
Flask-Migrate==2.5.3
Flask-SQLAlchemy==2.4.1
idna==2.9
itsdangerous==1.1.0
Jinja2==2.11.1
Mako==1.1.2
MarkupSafe==1.1.1
oauthlib==3.1.0
Pillow==7.0.0
PySocks==1.7.1
python-dateutil==2.8.1
python-dotenv==0.12.0
python-editor==1.0.4
requests==2.23.0
requests-oauthlib==1.3.0
six==1.14.0
SQLAlchemy==1.3.15
tweepy==3.8.0
urllib3==1.25.8
Werkzeug==1.0.0
